requirements.txt
uvicorn==0.20.0
fastapi==0.88.0
websockets==10.4
python-dotenv==0.21.0